+++ Sources/Point.h
@@ -10,31 +10,31 @@ namespace aga
     struct Point
     {
         Point ()
-          : X (0.0)
-          , Y (0.0)
+            : X (0.0)
+            , Y (0.0)
         {
         }
 
         Point (int x, int y)
-          : X (x)
-          , Y (y)
+            : X (x)
+            , Y (y)
         {
         }
 
-        Point (double x, double y)
-          : X (x)
-          , Y (y)
+        Point (float x, float y)
+            : X (x)
+            , Y (y)
         {
         }
 
         union {
             struct
             {
-                double X, Y;
+                float X, Y;
             };
             struct
             {
-                double Width, Height;
+                float Width, Height;
             };
         };
