+++ Sources/states/GamePlayState.h
@@ -24,9 +24,9 @@ namespace aga
         void BeforeEnter ();
         void AfterLeave ();
 
-        void ProcessEvent (ALLEGRO_EVENT* event, double deltaTime);
-        void Update (double deltaTime);
-        void Render (double deltaTime);
+        void ProcessEvent (ALLEGRO_EVENT* event, float deltaTime);
+        void Update (float deltaTime);
+        void Render (float deltaTime);
 
     private:
         Scene* m_CurrentScene;
